refactor(test): test 패키지 구조 리팩토링
어플리케이션 로직 패키지에 맞추어 test 패키지 구조 리팩토링
