+++ src/test/java/kitchenpos/order/domain/OrderMenuTest.java
@@ -1,6 +1,6 @@
 package kitchenpos.order.domain;
 
-import static kitchenpos.menu.domain.MenuGroupTestFixture.generateMenuGroup;
+import static kitchenpos.menugroup.domain.MenuGroupTestFixture.generateMenuGroup;
 import static kitchenpos.menu.domain.MenuProductTestFixture.generateMenuProduct;
 import static kitchenpos.menu.domain.MenuTestFixture.generateMenu;
 import static kitchenpos.order.domain.OrderMenuTestFixture.generateOrderMenu;
