+++ src/test/java/kitchenpos/ordertable/acceptance/TableRestAssured.java
@@ -1,4 +1,4 @@
-package kitchenpos.order.acceptance;
+package kitchenpos.ordertable.acceptance;
 
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
