+++ src/test/java/kitchenpos/ordertable/domain/OrderTablesTest.java
@@ -1,17 +1,15 @@
-package kitchenpos.order.domain;
+package kitchenpos.ordertable.domain;
 
 import static java.util.Collections.emptyList;
 import static java.util.Collections.singletonList;
-import static kitchenpos.order.domain.OrderTableTestFixture.generateOrderTable;
-import static kitchenpos.order.domain.TableGroupTestFixture.generateTableGroup;
+import static kitchenpos.ordertable.domain.OrderTableTestFixture.generateOrderTable;
+import static kitchenpos.tablegroup.domain.TableGroupTestFixture.generateTableGroup;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
 import java.util.Arrays;
 import kitchenpos.common.constant.ErrorCode;
-import kitchenpos.ordertable.domain.OrderTable;
-import kitchenpos.ordertable.domain.OrderTables;
 import kitchenpos.tablegroup.domain.TableGroup;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
